# [Enhancement] Optimize performance of arrays_overlaping