--- a/be/src/exprs/array_functions.h
+++ b/be/src/exprs/array_functions.h
@@ -51,6 +51,16 @@ public:
     }
 
     template <LogicalType type>
+    static Status array_overlap_prepare(FunctionContext* context, FunctionContext::FunctionStateScope scope) {
+        return ArrayOverlap<type>::prepare(context, scope);
+    }
+
+    template <LogicalType type>
+    static Status array_overlap_close(FunctionContext* context, FunctionContext::FunctionStateScope scope) {
+        return ArrayOverlap<type>::close(context, scope);
+    }
+
+    template <LogicalType type>
     static StatusOr<ColumnPtr> array_intersect(FunctionContext* context, const Columns& columns) {
         return ArrayIntersect<type>::process(context, columns);
     }