… gensrc/script/functions.py
… test/sql/test_array_fn/R/test_array_fn
… test/sql/test_array_fn/R/test_arrays_overlap
… test/sql/test_array_fn/T/test_array_fn
… test/sql/test_array_fn/T/test_arrays_overlap