## Why I'm doing:

`Datum` is not an efficient data structure, and `arrrasy_overlapping` heavily relies on Datum for processing, resulting in poor performance.

The second parameter of `arrrasy_overlapping` is usually const, but there is currently no special handling.

## What I'm doing:

PerformaceTest: 27.76s -> 2.74s

```
mysql> select count(*) from lineorder;
+-----------+
| count(*)  |
+-----------+
| 143999468 |
+-----------+
1 row in set (5.17 sec)

CREATE TABLE `t2` (
  `c1` bigint(20) NOT NULL COMMENT "",
  `c2` array<varchar(65533)> NOT NULL COMMENT "",
  `c3` array<varchar(65533)> NOT NULL COMMENT ""
) ENGINE=OLAP 
DUPLICATE KEY(`c1`)
DISTRIBUTED BY HASH(`c1`) BUCKETS 1 
PROPERTIES (
"compression" = "LZ4",
"fast_schema_evolution" = "true",
"replicated_storage" = "true",
"replication_num" = "1"
);

insert into t2 select lo_orderkey, array_agg(lo_partkey), array_agg(lo_suppkey) from lineorder group by lo_orderkey;

mysql> select count(*) from t2;
+----------+
| count(*) |
+----------+
| 30000000 |
+----------+
1 row in set (0.36 sec)
```

Before optimization: 27.76s

```
mysql> select count(*) from t2 where arrays_overlap(c2, ["1", "11", "111", "1111", "11111", "111111", "1111111"]);                                                                                                                                                                
+----------+                                                                                                                                                                                                                                                                      
| count(*) |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
|     1204 |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
1 row in set (27.76 sec)  
```

After optimization: 2.74s
```
mysql> select count(*) from t2 where arrays_overlap(c2, ["1", "11", "111", "1111", "11111", "111111", "1111111"]);                                                                                                                                                                
+----------+                                                                                                                                                                                                                                                                      
| count(*) |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
|     1204 |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
1 row in set (2.74 sec)
```

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5

## Documentation PRs only:

If you are submitting a PR that adds or changes English documentation and have not
included Chinese documentation, then you can check the box to request GPT to translate the
English doc to Chinese. Please ensure to uncheck the **Do not translate** box if translation is needed.
The workflow will generate a new PR with the Chinese translation after this PR is merged.

- [ ] Yes, translate English markdown files with GPT
- [x] Do not translate