@trueeyu 2024-09-30
https://github.com/Mergifyio backport branch-3.2