@trueeyu 2024-11-26
> now 3.3.5 crash *** Aborted at 1731327492 (unix time) try "date -d @1731327492" if you are using GNU date *** PC: @ 0x6c1cf10 starrocks::ArrayOverlap<(starrocks::LogicalType)17>::_check_column_overlap_nullable(phmap::flat_hash_set<starrocks::Slice, starrocks::PhmapDefaultHashFunc<(starrocks::LogicalType)17, (starrocks::PhmapSeed)0>, phmap::EqualTostarrocks::Slice, std::allocatox�� *** SIGSEGV (@0x0) received by PID 14351 (TID 0x7f73af977700) from PID 0; stack trace: *** @ 0x7f7af7cbbe20 __GI___pthread_once @ 0x7aff560 google::(anonymous namespace)::FailureSignalHandler(int, siginfo_t*, void*) @ 0x7f7af7cbe5e0 (/usr/lib64/libpthread-2.17.so+0xf5df) @ 0x6c1cf10 starrocks::ArrayOverlap<(starrocks::LogicalType)17>::_check_column_overlap_nullable(phmap::flat_hash_set<starrocks::Slice, starrocks::PhmapDefaultHashFunc<(starrocks::LogicalType)17, (starrocks::PhmapSeed)0>, phmap::EqualTostarrocks::Slice, std::allocatox�� @ 0x6c1d3d4 starrocks::ArrayOverlap<(starrocks::LogicalType)17>::_array_overlap_const(starrocks::ArrayOverlapState<phmap::flat_hash_set<starrocks::Slice, starrocks::PhmapDefaultHashFunc<(starrocks::LogicalType)17, (starrocks::PhmapSeed)0>, phmap::EqualTo<starrocks::Slx�� @ 0x6ce7d33 starrocks::StatusOr<std::shared_ptrstarrocks::Column > starrocks::ArrayFunctions::array_overlap<(starrocks::LogicalType)17>(starrocks::FunctionContext*, std::vector<std::shared_ptrstarrocks::Column, std::allocator<std::shared_ptrstarrocks::Column > >x�� @ 0x5d885c3 starrocks::VectorizedFunctionCallExpr::evaluate_checked(starrocks::ExprContext*, starrocks::Chunk*) @ 0x5c6a003 starrocks::VectorizedCastExpr<(starrocks::LogicalType)24, (starrocks::LogicalType)1, false>::evaluate_checked(starrocks::ExprContext*, starrocks::Chunk*) @ 0x5456df1 starrocks::VectorizedBinaryPredicate<(starrocks::LogicalType)1, starrocks::BinaryPredFunc<std::equal_to > >::evaluate_checked(starrocks::ExprContext*, starrocks::Chunk*) @ 0x525ba3b starrocks::ExprContext::evaluate(starrocks::Expr*, starrocks::Chunk*, unsigned char*) @ 0x525bebf starrocks::ExprContext::evaluate(starrocks::Chunk*, unsigned char*) @ 0x3f3fea3 starrocks::eager_prune_eval_conjuncts(std::vector<starrocks::ExprContext*, std::allocatorstarrocks::ExprContext* > const&, starrocks::Chunk*) @ 0x3f42646 starrocks::ExecNode::eval_conjuncts(std::vector<starrocks::ExprContext*, std::allocatorstarrocks::ExprContext* > const&, starrocks::Chunk*, std::shared_ptr<std::vector<unsigned char, std::allocator > >_, bool) @ 0x461cc6f starrocks::pipeline::OlapChunkSource::_read_chunk_from_storage(starrocks::RuntimeState_, starrocks::Chunk*) @ 0x461d171 starrocks::pipeline::OlapChunkSource::_read_chunk(starrocks::RuntimeState*, std::shared_ptrstarrocks::Chunk_) @ 0x4613478 starrocks::pipeline::ChunkSource::buffer_next_batch_chunks_blocking(starrocks::RuntimeState_, unsigned long, starrocks::workgroup::WorkGroup const*) @ 0x42e53e3 auto starrocks::pipeline::ScanOperator::_trigger_next_scan(starrocks::RuntimeState*, int)::{lambda(auto:1&)#1}::operator()starrocks::workgroup::YieldContext(starrocks::workgroup::YieldContext&) const [clone .constprop.0] @ 0x43e163b starrocks::workgroup::ScanExecutor::worker_thread() @ 0x38b3123 starrocks::ThreadPool::dispatch_thread() @ 0x38abb86 starrocks::Thread::supervise_thread(void*) @ 0x7f7af7cb6e25 start_thread @ 0x7f7af70bc34d __clone
> 
> mysql: SELECT t2.role_id as code, count(distinct t1.id) as count from ( select *, developers.unnest as developercode from view_dm_issue_detail, unnest (fieldkey_value_map['customfield_developers']) as developers where developers.unnest in ( '1111', '2222', '3333'
> 
> ```
>         )
>         AND PROJECT in (10244)
>         AND issuetype in (10201)
>         AND data_status = 1
>         AND arrays_overlap(
>             IFNULL (fieldkey_value_map['customfield_bug_source'], []),
>             ['开发', '测试, '产品']
>         ) = 1
>         AND arrays_overlap(
>             IFNULL (fieldkey_value_map['customfield_iteration'], []),
>             [
>                 '1',
>                 '2',
>                 '3'
>             ]
>         ) = 1
> ) t1
> join view_dm_project_role_user t2 on t1.developercode = t2.usercode
> ```
> 
> where t2.project_id = 10244 group by t2.role_id;

Fixed by this pr: https://github.com/StarRocks/starrocks/pull/52866